Removed DataUriSuffixParser and updated DataUriParser.
