+++ app/src/androidTest/java/com/duckduckgo/app/browser/downloader/DataUriParserTest.kt
@@ -28,7 +28,7 @@ class DataUriParserTest {
 
     @Before
     fun setup() {
-        testee = DataUriParser(DataUriSuffixParser())
+        testee = DataUriParser()
     }
 
     @Test
@@ -71,5 +71,59 @@ class DataUriParserTest {
     fun whenMimeTypeNotProvidedThenNoSuffixAddedInToString() {
         val filename = testee.generate("data:,AAAA") as ParsedDataUri
         assertFalse(filename.toString().contains("."))
+    }
+
+    @Test
+    fun whenMimeTypeIsTextPlainAndDataIsBase64AndIsPngThenSuffixIsPng() {
+        val parsed = testee.generate("data:text/plain;base64,iVBORw0KGgo") as ParsedDataUri
+        assertEquals("png", parsed.filename.fileType)
+    }
+
+    @Test
+    fun whenMimeTypeIsTextPlainAndDataIsBase64AndIsJfifThenSuffixIsJpg() {
+        val parsed = testee.generate("data:text/plain;base64,/9j/4AAQSkZJRgABAQ") as ParsedDataUri
+        assertEquals("jpg", parsed.filename.fileType)
+    }
+
+    @Test
+    fun whenMimeTypeIsTextPlainAndDataIsBase64AndIsSvgThenSuffixIsSvg() {
+        val parsed = testee.generate("data:text/plain;base64,PHN2ZyB2ZXJzaW9uPSIxLjIiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy") as ParsedDataUri
+        assertEquals("svg", parsed.filename.fileType)
+    }
+
+    @Test
+    fun whenMimeTypeIsTextPlainAndDataIsBase64AndIsGifThenSuffixIsGif() {
+        val parsed = testee.generate("data:text/plain;base64,R0lGODlhAAXQAocAAP") as ParsedDataUri
+        assertEquals("gif", parsed.filename.fileType)
+    }
+
+    @Test
+    fun whenMimeTypeIsTextPlainAndDataIsBase64AndIsPdfThenSuffixIsPdf() {
+        val parsed = testee.generate("data:text/plain;base64,JVBERi0xLjEKMSAwIG9iag") as ParsedDataUri
+        assertEquals("pdf", parsed.filename.fileType)
+    }
+
+    @Test
+    fun whenMimeTypeIsTextPlainAndDataIsBase64AndIsWebpThenSuffixIsWebp() {
+        val parsed = testee.generate("data:text/plain;base64,UklGRs4IAABXRUJQVlA4WAo") as ParsedDataUri
+        assertEquals("webp", parsed.filename.fileType)
+    }
+
+    @Test
+    fun whenMimeTypeIsTextPlainAndDataIsBase64AndIsBpThenSuffixIsBmp() {
+        val parsed = testee.generate("data:text/plain;base64,Qk1AwgEA") as ParsedDataUri
+        assertEquals("bmp", parsed.filename.fileType)
+    }
+
+    @Test
+    fun whenMimeTypeIsTextPlainAndDataIsBase64AndIsUnknownThenSuffixIsTxt() {
+        val parsed = testee.generate("data:text/plain;base64,RUJQVlA4WAo") as ParsedDataUri
+        assertEquals("txt", parsed.filename.fileType)
+    }
+
+    @Test
+    fun whenMimeTypeIsImageJpegThenSuffixIsJpg() {
+        val parsed = testee.generate("data:image/jpeg;base64,RUJQVlA4WAo") as ParsedDataUri
+        assertEquals("jpg", parsed.filename.fileType)
     }
 }
